Adding modal boxes together to map
