+++ src/components/index.jsx
@@ -8,3 +8,4 @@ export { default as JourneyPlanCard } from './JourneyPlanCard';
 export { default as DirectionsMap } from './DirectionsMap';
 export { default as Notification } from './Notification';
 export { default as GetNotifications } from './GetNotifications';
+export { default as LocationModal } from './LocationModal';
